services/detector/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
ultralytics>=8.0.0
pillow>=10.0.0
pydantic>=2.0.0
python-multipart>=0.0.6
